🚚 move recover modal as a component of the recoverpassword component
